+++ Ddugyesangi/Views/ListRowView.swift
@@ -44,7 +44,6 @@ struct ListRowView: View {
             case .project(let projectListViewModel):
                 if let project = project {
                     ProjectEditView(project: project, viewModel: projectListViewModel, isPresented: $showingEditSheet)
-                        .presentationDetents([.fraction(0.25)])
                 }
             case .part(let partListViewModel):
                 if let part = part {