+++ Ddugyesangi/Views/Project/ProjectEditView.swift
@@ -27,8 +27,9 @@ struct ProjectEditView: View {
         NavigationStack {
             VStack(spacing: 20) {
                 NomalTextField(placeholder: NSLocalizedString("Project Name", comment: ""), text: $projectName)
-                Spacer()
             }
+            .padding(.vertical)
+            .fitSheetToContent()
             .navigationTitle("Edit Project")
             .navigationBarTitleDisplayMode(.inline)
             .toolbar {